+++ projects/batfish/src/test/java/org/batfish/grammar/cisco_nxos/CiscoNxosGrammarTest.java
@@ -6442,14 +6442,14 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.0.0/24")).iterator().next();
       assertTrue(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
     }
     {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.1.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
     }
@@ -6457,7 +6457,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.2.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6466,7 +6466,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.13.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), nullValue());
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6475,7 +6475,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.3.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6485,7 +6485,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.4.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6496,7 +6496,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.5.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(1000L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6507,7 +6507,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.6.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 5));
+      assertThat(route.getPreference(), equalTo(5));
       assertThat(route.getTag(), equalTo(1000L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6518,7 +6518,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.7.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 5));
+      assertThat(route.getPreference(), equalTo(5));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6529,7 +6529,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.8.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 5));
+      assertThat(route.getPreference(), equalTo(5));
       assertThat(route.getTag(), equalTo(1000L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.1.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/1"));
@@ -6540,7 +6540,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.9.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 5));
+      assertThat(route.getPreference(), equalTo(5));
       assertThat(route.getTag(), equalTo(1000L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.2.254")));
       assertThat(route.getNextHopInterface(), nullValue());
@@ -6552,7 +6552,7 @@ public final class CiscoNxosGrammarTest {
       StaticRoute route =
           vc.getDefaultVrf().getStaticRoutes().get(Prefix.strict("10.0.10.0/24")).iterator().next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 5));
+      assertThat(route.getPreference(), equalTo(5));
       assertThat(route.getTag(), equalTo(1000L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.2.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/2"));
@@ -6569,7 +6569,7 @@ public final class CiscoNxosGrammarTest {
               .iterator()
               .next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.2.254")));
       assertThat(route.getNextHopVrf(), nullValue());
@@ -6583,7 +6583,7 @@ public final class CiscoNxosGrammarTest {
               .iterator()
               .next();
       assertFalse(route.getDiscard());
-      assertThat(route.getPreference(), equalTo((short) 1));
+      assertThat(route.getPreference(), equalTo(1));
       assertThat(route.getTag(), equalTo(0L));
       assertThat(route.getNextHopIp(), equalTo(Ip.parse("10.255.2.254")));
       assertThat(route.getNextHopInterface(), equalTo("Ethernet1/100"));
